+++ R/utils-collect.R
@@ -1,0 +1,10 @@
+#' Collect function
+#'
+#' See \code{dplyr::\link[dplyr]{collect}} for details.
+#'
+#' @name collect
+#' @rdname collect
+#' @keywords internal
+#' @export
+#' @importFrom dplyr collect
+NULL